Move rated cocktails list to Favorite screen
